refactor for middleware
